LPS-54970 move SassToCssBuilder to module with a more appropriate name as the class does much more than convert sass to css.
