+++ modules/util/portal-tools-css-builder/src/com/liferay/portal/tools/css/builder/sass/SassFileWithMediaQuery.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.tools.sass;
+package com.liferay.portal.tools.css.builder.sass;
 
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
